chore: move bond calculating logic to bonds feature
